Change requirement from "pexpect<2.5" to "pexpect"
In Fedora18, we don't want pexpect version 2.5 since it's
incompatible with python2.7. That's the reason why requirement
in spec file was "pexpect<2.5".

But version 2.5 is provided by package named "python-pexpect"
and there is a package pexpect-2.3 which is suitable for us.

Error occurs if don't change this requirement:
---> Package itest-core.noarch 0:1.3-2.1 will be installed
--> Processing Dependency: pexpect < 2.5 for package: itest-core-1.3-2.1.noarch
Package pexpect is obsoleted by python-pexpect, but obsoleting package does not provide for requirements
--> Finished Dependency Resolution
Error: Package: itest-core-1.3-2.1.noarch (tools)
	   Requires: pexpect < 2.5
	   Available: pexpect-2.3-8.fc18.noarch (fedora)
	       pexpect = 2.3-8.fc18
	   Available: python-pexpect-2.5.1-5.fc18.noarch (fedora)
	       pexpect = 2.5.1-5.fc18
 You could try using --skip-broken to work around the problem
 You could try running: rpm -Va --nofiles --nodigest

Change-Id: Ic313f0de160defaf0e9b111b78a6ada8df3fb85c