Skip a test when a SkipTest exception is raised.
Rearrange the code of TestCase.run to deal with the SkipTest
exception, which can skip current test in some conditions.

Call result.test_stop when test already had finished, which is
the same as unittest stdlib. As a consequence, information should
be print in TestResult.test_stop() rather than TestResult.add_*().

Change-Id: I8c1f95f4e7d97cbf94acae4ba440260cd5f47221